fix: fixing linux ci issues, adding -Wno-changes-meaning because who cares

+++ src/PhoenixSim/FixedPoint/FixedVector.h
@@ -321,10 +321,10 @@ namespace Phoenix
     using Vec3 = TVec3<Distance>;
     
     template <class> struct TZero {};
-    template <uint32 Td, class T> struct TZero<TFixed<Td, T>> { static constexpr TFixed<Td, T> Value = 0; };
+    template <uint8 Td, class T> struct TZero<TFixed<Td, T>> { static constexpr TFixed<Td, T> Value = 0; };
     template <class T> struct TZero<TVec2<T>> { static constexpr TVec2<T> Value = TVec2<T>(0, 0); };
 
     template <class> struct TOne {};
-    template <uint32 Td, class T> struct TOne<TFixed<Td, T>> { static constexpr TFixed<Td, T> Value = 1; };
+    template <uint8 Td, class T> struct TOne<TFixed<Td, T>> { static constexpr TFixed<Td, T> Value = 1; };
     template <class T> struct TOne<TVec2<T>> { static constexpr TVec2<T> Value = TVec2<T>(1, 1); };
 }
